+++ icicle/tests/test_ring_api.cpp
@@ -1,7 +1,29 @@
 
 
-#include "test_ring_field.h"
+#include "test_mod_arithmetic_api.h"
 
-// Add ring specific tests here
+// Derive all ModArith tests and add ring specific tests here
+template <typename T>
+class RingTest : public ModArithTest<T>
+{
+};
 
-// TYPED_TEST(RingAndFieldTest, /*Test Name*/) {}
+TYPED_TEST_SUITE(RingTest, FTImplementations);
+
+// Note: this is testing host arithmetic. Other tests against CPU backend should guarantee correct device arithmetic too
+TYPED_TEST(RingTest, RingSanityTest)
+{
+  auto a = TypeParam::rand_host();
+  auto b = TypeParam::rand_host();
+  auto a_neg = TypeParam::neg(a);
+  ASSERT_EQ(a + TypeParam::zero(), a);
+  ASSERT_EQ(a + b - a, b);
+  ASSERT_EQ(a + a_neg, TypeParam::zero());
+  ASSERT_EQ(a * TypeParam::zero(), TypeParam::zero());
+  ASSERT_EQ(a * scalar_t::from(2), a + a);
+
+  // TODO Yuval: ring inverse
+  //   auto b_inv = TypeParam::inverse(b);
+  //   ASSERT_EQ(b * a * b_inv, a);
+  //   ASSERT_EQ(b * b_inv, TypeParam::one());
+}
